Breaks and Racket
The reset position button is now the right stick button. Pressing the A
button now adds a lot of drag to the ball causing it to slow down and
stop quickly, releasing the a button resets the drag back to normal. I
also created a racket like object for the right hand controller