Added MirrorCube and Build for mid-semester meeting
- The build is stored locally but not commited to Github because of
.gitignore
- Torus scene: move the ball farther away from table to avoid physical
damage. It is recommended to tell the player to stand at the center of
the working area of Oculus Rift.
- MirrorCube isn't functioning perfectly but playable.
- Duplicator.cs and Wrapper.cs: changed the accessibility of attributes
and methods from 'private' to 'protected' to allow inheritance.
- ManualBouncePhysics.cs: check if rigidBody is null because sometimes
Wrapper doesn't work (?!) and stops detecting the closest ball. If
player tries to hit that ball, an error will appear because that ball
doesn't have a rigidbody.
- Changed program name in ProjectSettings