Initiated Grab feature
The grab feature is implemented in PhysicsUtil.cs but not working
correctly after a throw. Basically, when the ball is released from the
glove, the Y value of its positional vector turns to -1 (below the desk)
for an unknown reason. That value changed even when I was throwing the
ball horizontally. Surprisingly, this problem doesn't happen when I
throw vertically.