Create Tiling class to unify wrap and clone logic
Now all calculations related to the manifold are in class Tiling.

This class offers two services:

1) It is IEnumerable, so you can loop over it,

  foreach (T in tiling) { ... }

will iterate over 4x4 matrices T that move the unit cell to each other
cell of the tiling.

2) It has a method TileContaining(P) which takes a 4x4 matrix P and
returns the transformation that should be applied to pose P in order to
bring it back to the unit cell.

Now Registrar maintains the single global instance of Tiling that
everyone else uses; this instance is constructed with N and cubsize from
the public attributes of Registrar.

Registrar uses tiling to make the clones.
Wrapper pulls the tiling instance from Registrar, and then uses it for
the wrapping calculations.

MISSING for now is the old feature to make sure that the unit cell is
always centered at the ball's initial position.  That will require a
modification to the constructor of the Tiling class.