Restore menu, duplicate main scene w/ 3 variants
This commit introduces three nearly-identical scenes which were copied
from TorusGame.  They differ only in the initial value of "TilingType" on
the Registrar script.

TODO: Refactor so that we don't need to maintain three copies of the
scene.  Instead, there should be a single game scene which the MenuScene
passes a parameter indicating which tiling to use.